requirements.txt
allure-pytest
allure-python-commons
httpx==0.22.0
playwright==1.21.0
pytest==7.1.1
pytest-asyncio==0.18.3
pylint-fail-under
pytest-xdist
pytest-bdd~=5.0.0
pytest-factoryboy
pytest-rerunfailures
playwright-stealth
pytest-playwright==0.3.0
pytest-playwright-snapshot
pytest-html
pytest-html-reporter==0.2.9
pytest-metadata==2.0.1
pytest-order==1.0.1
simplejson~=3.17.6
pathlib2==2.3.7.post1
py~=1.11.0
pytest-reportportal==5.1.0
pyleniumio==1.15.3
Faker==13.3.4
pytest-json-report==1.5.0
PyHamcrest
rich==12.2.0